Renaming interfaces and namespaces
