+++ DSL.ReqnrollPlugin/Matches/UserVariableMatchInterpreter.cs
@@ -1,6 +1,6 @@
 ﻿using System.Text.RegularExpressions;
 
-namespace DSL.ReqnrollPlugin
+namespace DSL.ReqnrollPlugin.Matches
 {
     public static class UserVariableMatchInterpreter
     {
