+++ DSL.ReqnrollPlugin/ReqnrollPluginTestRunner.cs
@@ -1,4 +1,5 @@
 using System;
+using DSL.ReqnrollPlugin.Transformers;
 using Reqnroll;
 using Reqnroll.Infrastructure;
