+++ DSL.ReqnrollPlugin/Transformers/EnvironmentVariableTransformer.cs
@@ -1,12 +1,13 @@
-﻿using Reqnroll;
+﻿using DSL.ReqnrollPlugin.Matches;
+using Reqnroll;
 using System;
 
-namespace DSL.ReqnrollPlugin
+namespace DSL.ReqnrollPlugin.Transformers
 {
     public class EnvironmentVariableTransformer : BaseParameterTransformer, IEnvironmentVariableTransformer
     {
         protected override string TransformText(in string inputString, in ScenarioContext scenarioContext) => TransformText(inputString);
-        
+
         public virtual string GetEnvironmentVariable(string key) => key == null ? key : Environment.GetEnvironmentVariable(key);
 
         public virtual string TransformText(in string inputString)
@@ -16,8 +17,8 @@ namespace DSL.ReqnrollPlugin
             var match = PatternMatch.Parse(inputString, PatternMatchConfig.EnvironmentMatchConfig);
             var envVariableValue = GetEnvironmentVariable(match?.MatchedPattern);
 
-            return (match == null || string.IsNullOrEmpty(envVariableValue)) 
-                ? inputString 
+            return match == null || string.IsNullOrEmpty(envVariableValue)
+                ? inputString
                 : TransformText(match.ReplaceMatched(envVariableValue));
         }
     }
