--- DSL.ReqnrollPlugin/Transformers/IParameterTransformer.cs
@@ -1,10 +1,0 @@
-﻿using System;
-
-namespace DSL.ReqnrollPlugin
-{
-    public interface IParameterTransformer : ITransformer
-    {
-        IParameterTransformer AddBespokeTransformer(in Func<string, string> transformer);
-        void ClearBespokeTransformers();
-    }
-}
