+++ DSL.ReqnrollPlugin/Transformers/ITransformer.cs
@@ -1,6 +1,6 @@
 ﻿using Reqnroll;
 
-namespace DSL.ReqnrollPlugin
+namespace DSL.ReqnrollPlugin.Transformers
 {
     public interface ITransformer
     {
