+++ DSL.ReqnrollPlugin/Transformers/IUserVariableTransformer.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace DSL.ReqnrollPlugin.Transformers
+{
+    public interface IUserVariableTransformer : ITransformer
+    {
+        IUserVariableTransformer AddBespokeTransformer(in Func<string, string> transformer);
+        void ClearBespokeTransformers();
+    }
+}
